renamed the dycore subpackage to flow_solver for consistency
