+++ src/flow_solver/utils/variable.py
